refactor: move NoteDecoder and StreamDisplay to their own files
